+++ tests/testthat/test_bow.R
@@ -1,6 +1,4 @@
 context("Bow")
 library(polite)
 
-test_that("Bow warns about disallowed url", {
- expect_warning(bow("https://www.wikipedia.org/w/", verbose = TRUE), "scrape")
-})
+