+++ tests/testthat/test_nod.R
@@ -1,8 +1,4 @@
 context("Nod")
 library(polite)
 
-test_that("Nod warns about disallowed url", {
-  expect_warning(
-    nod(bow("https://www.wikipedia.org/"),
-        path="w/", verbose = TRUE), "scrape")
-})
+